Pavel/chip-list abstract (#1492)
* selectMode classic for tree list

* chip-list abstract
