+++ projects/ui-framework/src/lib/chips/chip-list/chip-list.module.ts
@@ -3,12 +3,11 @@ import { CommonModule } from '@angular/common';
 import { ChipListComponent } from './chip-list.component';
 import { ChipModule } from '../chip/chip.module';
 import { AvatarModule } from '../../avatar/avatar/avatar.module';
-import { EventManagerPlugins } from '../../services/utils/eventManager.plugins';
 
 @NgModule({
   declarations: [ChipListComponent],
   imports: [CommonModule, ChipModule, AvatarModule],
   exports: [ChipListComponent],
-  providers: [EventManagerPlugins[0]]
+  providers: [],
 })
 export class ChipListModule {}
